refactor: 불필요한 메서드 제거 및 패키지 이동
