+++ mvc/src/main/java/nextstep/mvc/controller/AnnotationHandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.lang.reflect.Method;
