+++ mvc/src/main/java/nextstep/mvc/controller/ControllerScanner.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.controller;
 
 import static org.reflections.Reflections.log;
